2. Расширенное хранилище.

+++ chapter_007/src/main/java/ood/lsp/foodstorage/products/Food.java
@@ -60,7 +60,7 @@ public class Food implements Product {
     }
 
     @Override
-    public double getDiscount() {
+    public int getDiscount() {
         return discount;
     }
